Factor CoNLL-2003 downloading code out of util.py
