+++ requirements.txt
@@ -5,4 +5,4 @@ regex
 fastparquet
 # TODO: The following dependency should go away when we switch to Python 3.8.
 memoized-property
-
+requests
